Variables: Fixes issue with upgrading legacy queries
